re=structured some other folders
