--- mimic/gMLV/__init__.py
@@ -1,3 +1,0 @@
-# __init__.py
-from ..model_infer.gMLV_ML import *
-from .multi_penalty_lasso import *
