+++ mimic/model_infer/__init__.py
@@ -3,5 +3,8 @@
 
 # Import the VAR class from the VAR module
 from .VARinfer import *
+from .gMLV_ML import *
+from .multi_penalty_lasso import *
+from .cLV import *
 
 # You can also include some package-level documentation
